I really need that night's sleep. I've been tallying my wins and loses during the night. Posting on the ML sub was definitely a loss, but the probe did serve its purpose.
Posting on the Haskell sub was a surprising win and if I want to spread Spiral's word, I should focus on that particular crowd. Those guys are surprisingly open minded, especially to upvote posts like that in an argument with ekmett that I initiated.

Yesterday was really rough on me. I blew my creative load on making those posts, and ended running on fumes for the rest of the day. I still got some work, but I am going to have to make up for the slacking that I missed today.

Since I've received no replies today, I think I will go to my original plan of slipping into the shadows and submerging myself into my work.

First let me make a little outline of what I want to do.

My like in martial arts, the main focus should be on mastering the basic punch, so will that be my goal with regards to my ML library.

There are great many things Spiral will allow me to do that other languages won't, but that doesn't really matter for now.

My main goal is to make portal from which I can draw on a steady amount of good gameplaying agents.

Without a doubt, there is still much research to be done before the misteries of neural nets can be unravelled, and I am going to have to use Spiral to bring out not just their compositionality which I am doing now, but also their abstractive abilitity - but that is not my focus right now.

What I've been trying to do for the past 2.5 years and failing is the most basic of the basics.

I have a little prediction I want to try out. What the current crop of NNs represent are focus - insane, unwavering focus to the point of leaving everything else aside. They do not have any ability for abstraction or reasoning.

In my mental experiments, I think that the neural nets actually capable of human like reasoning will actually get a lot lower scores on various benchmarks than the nets that exist today.

Power. What the current crop of neural nets represent are pure, unbridled power.

This is good.

And this is exactly what I need.

My prediction is that the current state of the art is just what I need to master poker. I do not need much else.

But in order to get the necessary margin of safety I need to do it and do it with ease. In order to bring out their compositionality and create the necessary foundations for when I am done with poker, I need to make this upfront investment.

The smallish library I have now might now seem much, but I am throwing a lot of power around at this point with very little code.

Edit: Done chilling, let me start.