Yeah, I am done for the day. It it time to start recharging.
The two new kernels are not done by any stretch I am sure they aren't even correct right now.

But...this is real programming.

Before I started this, I was actually a bit skeptical of whether it will be possible, but then it absolutely hit me. When I can to face the dragon, the full force of my focus came onto me, uplifting my spirit.

That familiar feeling of power, of pieces falling into place is why I created Spiral.

It is possible, it is possible to go much deeper into the secrets of programming with Spiral, this is not like programing in raw C when I had to keep everything in my head.

I can reason properly about the code now.

The only thing left would be to bring it to a close.

I will dedicate tomorrow to getting these kernels done, but spending as much time as needed in order to internalize the loops would not be bad either, so I won't set any deadlines.

These kernels will become the backbone of Spiral's ML library and learning new ways of reasoning will pay off greatly down the line even for architectures other than CPUs.