The two kernels work. They literally worked fine the first time after they compiled without error.
I have some TODO's left, but I am going to leave that for during the performance tuning.

Let me take a break here.

Things went so smoothly today that it is absolutely unbelievable. I am still not done yet of course. It is time that I put this stuff into action and implement the biases. After that I will look into performance tuning.

Just now I was looking at the generated PTX and have noticed the the compiler automatically untrolled all the loops that it could. I actually forgot about the shared memory loop.

This leds credence to the idea that I should try my best to propagate as much size information through the system. In the real world I do not need that last minibatch if it complicates compilation.

It won't be long - it won't be long at all until I've caught up to my previous library. And it won't be long until I significantly exceed either it or the competing language.