Shit I did the replicate map incorrectly. And the d2_redo_map. I have no way of throwing out the unneeded primal here.
In replicate_map I need an extra input that does not need to be replicated.

I won't if I can take advantage of the empty tensor for that?

Ugh, forget this for a bit. Instead of just testing the kernel, I need to actually test for adding biases.

Let me see if I can manage to get this done today.