Is the Github up?
At any rate, the first version of the kernels was incomplete so I had to scramble a little.

The change did involve some design work, but not to the way the loop works. That has remained the same.

Fuuu, damn I am tired.

Design is not an easy thing. If programmign was merely about solving puzzles the Singularity would have already happened.

Tomorrow I need to do more tensting of these functions as there is stuff not on the hot path now, but it is not a big deal.

It probably works fine either way.

I think tomorrow I will be able to give making AD versions of them a crack again. If everything goes according to plan I will be done with the biases and will have moved on to performance tuning.

After that...I guess what will come is accuracy. I need a kernel for that as well.

And after that, I will try stacking layers and such.

After that, it will time to stop laying around and get back to where I was last year when I threw in towel.

This time there will be no failure. With the power of the Spiral language, nothing will stop me - nothing.

Maybe by the end of February I will have finished the buildups for good and can working on poker players.