Let me take a break here.
I thought the biases would lead to better performance, but as it turns out it is not much.

Actually, instead of moving to tuning what I will do is put in the accuracy. That should be the next step.

Only after that will I start the build-up.

I won't be long now. I am motivated to do this and in a few weeks I will have more than caught up to where I was last year. After that it is poker time.