Ok, it turns out the cost was merely not being printed correctly. That has been fixed.
Honestly, I'd rather call it a day now, but let me get that thing out of the way.

...No, I can't think like this. Really making a new kernel function is really a huge effort and I can't do it unless I am fresh.

Let me compromise. Let me just hack add_bias so it works in place. That should seal the deal for today. Doing that will be much easier and will free up more time for later.