"5:40pm.     inl replicate_map =         inl body is_inplace {fwd bck={bck_in bck_in'}} in in' ret =             inl primal, adjoint = primals in, adjoints in             inl primal', adjoint' = primals in', adjoints in'             inb out =                 if is_inplace then inl ret -> replicate_map' (inl a b _ -> fwd a b) primal primal' primal'; ret in'                 else replicate_map fwd primal primal' >>! dr             ret (out, inl _ ->                 inl out = match out with {DR={primal adjoint}} -> zip (primal, adjoint) .update_body2 (inl P A -> {P A})                 on_non_nil adjoint (map_d2_redo_map' bck_in {in'=primal'; out} primal)                 on_non_nil adjoint' (replicate_map' bck_in' primal {in'=primal'; out})                 )         function         | .inplace -> body true         | x -> body false x I finally got it to compile properly. I think instead of primal' I should have return in' instead. My bad.
I won't worry about out here. I'll just assume that the dual will be plugged in. Forget everything else. Let me see if this works.
-----
Batch done.
Average of batch costs is 0.724580331389109.
-----
-----
Batch done.
Average of batch costs is 0.717539751116435.
-----
-----
Batch done.
Average of batch costs is 0.711618011283874.
-----
-----
Batch done.
Average of batch costs is 0.706528938579559.

Sure looks like it.

Let me compare.

-----
Batch done.
Average of batch costs is 0.751134147453308.
-----
-----
Batch done.
Average of batch costs is 0.74301914580663.
-----
-----
Batch done.
Average of batch costs is 0.736049199008942.
-----
-----
Batch done.
Average of batch costs is 0.729927414798737.
-----
Yeah, it is not right unfortunately.

5:50pm. Yeah, I can't accept this after all. I have no idea what the problem is though, but this much could have been expected from mutation.

The whole reason why I am testing this is because even as I finished the function, I could feel my grasp on it slipping. I think instead of trying to figure out what is wrong, I will redesign it so I can be sure of what it right...

Probably what is going on though is the some of the adjoints are getting added twice.

5:55pm. Let me make `matmultb`."