Yeah, unlike that mutable piece of crap, matmultb works properly. This shoudl be a lesson to me.
Now...yeah I will call it a day here.

I had enough. I definitely put in enough today.

Tomorrow I will make the accuracy kernel and integrate it with the cost functions.

The real reason this is going so slowly is not because I am slow, but because nobody bothered to make Spiral and do this years ago.

Well, let me chill, maybe I will finish that Tensorflow CLF video. I got bored third of the way in and started programming instead today.