I got up early today and spent all this time composing posts on Reddit.
I thought I was done with the replies for the time being, but Reddit somehow dropped the replies and I did not see them. Well, one was from two days ago and the other from 4, so I did not miss them by too much.

Let me have breakfast and then I will start.

Though it took me a while, it did not kill me like the day after I posted Spiral on the ML sub, so I'll find time to deal with the kernel that needs to be done today.

For now breakfast.