Let me start the day. I got up 1.5 hours ago and I've spent most of that time crafting a Reddit post. Maybe I should not have posted the language on the ML sub given how much time it is sucking out of me.
I think that this will be the last time.

Before I start, I need to make a little review of wins and losses from yesterday.

I am embarassed. Despite using the distractions as an excuse, I am not doing a good job.

The optimized LSTM cell that keeps coming into my mind is trying to tell me a story.

That story is that it should be the common sense.

https://devblogs.nvidia.com/parallelforall/optimizing-recurrent-neural-networks-cudnn-5/

When I saw it last year, I had the very real indication that where I was going with the previous library wasn't anywhere I really wanted to go.

Without a doubt that article made an impact on me.

I want to use RNNs. I want to use RNNs to do reinforcement learning. It is the whole reason why I made Spiral.

Since those guys got a 11x speedup, this is not something I can leave lying on the table. I need to pick that thing up.

There is also something that has been bothering me for a while - the lazy values in cost functions.

I know why I put them into the previous library, but the kicker is - when doing concurrent computation, batching it is pretty much always superior to launching multiple kernels.

So for example for calculating multiple costs, I'd be better off doing them in one kernel. Same fusing pointwise operators.

For concurrent gemms, I am going to use the batched gemm in the library, no way around that. Or maybe write my own gemm depending on the structure of the work.

Since I am going to have to do huge transformations in order to batch all the operations in order to get them up to speed, I should just simplify my life and get rid of the unnecessary things like lazy values.

That optimized LSTM will not be a huge thing to cause an upset, those are just my dumb daydreams.

Rather it will serve as the main benchmark of the new library for all kinds of architectures.

It will be the new standard.

The thing in that post will be the commons to strive for. It won't be a big deal.

Now with that out of the way, let me get breakfast since I wasted my morning yet again today.

After that I willl figure out what is causing the compilation errors on the Cuda side and investigate the bloat along with testing the accuracy.

I am really motivated today in contrast to yesterday. I will get this done.