Ah, now what do I do about this? Well, I know what the error is, but what do I do about it?
I've done something really stupid last time in the codegen, but the general intention of it is sound.

There is no way around it, deal with this will take me a while. Let me get through it. I am going to remove the nested arrays from buffer_type_definition and try introducing a tier somehow.