"1:15pm. Ok, it can't possible be in map_d1_redo_map. It has to be somewhere else.     inl map_dx_redo_map_template dim kernel d in in' ret =         inl in' =             match in' with             | () -> HostTensor.create {elem_type=(); dim}             | in' -> zip in'
        inl map_in = match d with {map_in} -> map_in | _ -> const
        inl map_out, elem_type =
            inl ty = type map_in in.elem_type in'.elem_type
            match d with {map_out} -> (inl a _ -> map_out a),(type map_out ty) | _ -> const, ty
        inb out = create {elem_type dim=in'.dim}
        kernel {d with map_in map_out} in in' out
        ret out

    inl map_d1_redo_map d (!zip in) = map_dx_redo_map_template (fst in.dim) map_d1_redo_map' d in

It has to be here then, but...
    inl accuracy (input,label) ret =
        inl input, label = primal input, primal label
        //inb x =
        //    map_d1_redo_map {
        //        map_in=const
        //        neutral_elem=-infinity,zero
        //        redo=inl a b -> if fst a > fst b then a else b
        //        map_out=inl a -> snd a = one
        //        } (input,label) ()
        //ret (Array.foldl (inl s x -> if x then s+1 else s) 0 (to_host_tensor x).bodies.ar)
        ret 0

Yeah, it is definitely somewhere in this commented out block of code.
        inb x =
            map_d1_redo_map {
                map_in=const
                neutral_elem=-infinity,zero
                redo=inl a b -> if fst a > fst b then a else b
                map_out=inl a -> snd a = one
                } (input,label) ()

Uncommenting this out does not break it.
//ret (Array.foldl (inl s x -> if x then s+1 else s) 0 (to_host_tensor x).bodies.ar)

So the error is here?

Is it in to_host_tensor x?

inl _ = to_host_tensor x
Yeah, it is. What gives?

The network is definitely getting specialized twice for some reason. I'll figure out why later.

But now why is to_host_tensor giving me a problem?
[|False; False; False; False; False; False; False; False; False; False; False; False; False; True; False; False; False; False; False; False; False; False; False; True; False; False; False; False; True; False; False; False|]

So it actually prints. Interesting. I'd expected it to crap the first thing.

1:30pm. I have no idea. Let me try accuracy right there in place.

1:35pm. There is a thought that is trickling through my mind. It occurs to me that this is the first time I really used boolean arrays.

What would happen if I use int64 arrays and summed those up instead?

Unhandled Exception: ManagedCuda.CudaException: ErrorMisalignedAddress: While executing a kernel, the device encountered a load or store instruction on a memory address which is not aligned.
The context cannot be used, so it must be destroyed (and a new one should be created).
All existing device memory allocations from this context are invalid and must be reconstructed if the program is to continue using CUDA.
   at ManagedCuda.CudaContext.CopyToHost(Int64[] aDest, CUdeviceptr aSource) in i:\ManagedCuda\managedCuda\ManagedCUDA\CudaContext.cs:line 3292
   at SpiralExample.Main.method_15(CudaContext var_0, CUmodule var_1, CudaStream var_2, UInt64 var_3, UInt64 var_4, Stack`1 var_5, FSharpRef`1 var_6, FSharpRef`1 var_7, CudaBlasHandle var_8, FSharpRef`1 var_9, FSharpRef`1 var_10, FSharpRef`1 var_11, FSharpRef`1 var_12, Int64 var_13, Double var_14, Int64 var_15) in C:\Users\Marko\Source\Repos\The Spiral Language\Temporary\output.fs:line 1022

Now this is new. Just what is going on here?

https://devtalk.nvidia.com/default/topic/975906/cuda-runtime-api-error-74-misaligned-address/?offset=5
https://stackoverflow.com/questions/12778949/cuda-memory-alignment

Ok, I could understand the char arrays if that was the error, but what is this? Why I am getting these errors when I am using int64 arrays?

Is Cub doing something stupid with tuples?"

Memory errors are the worst. Let me push onward.