I must have wondered into hell here somehow. Chars outright corrupt the stack, float32s give incorrect results and int64 give misalligned memory accesses.
What?

It feels like not just one bug, but several at once.

Also it turns out there was a huge bug in to_dev_tensor where I forgot to add the type size to the offsets.

I was hoping that would at least fix the misalligned accesses, but I have no such lunch.

For from being a simple kernel, map_d1_redo_map is turning out to be the most difficult of the bunch.

I am completely dry and am going to have to make a specialized test just for this particular thing.

First of all, let me try training on a single minibatch and seeing how the accuracy responds. After that I will try printing it to make sure whether it is overfitting as expected or if accuracy is by some miracle really correct.

Let me take a short break first and I will continue after that.