I was right that there are multiple bugs in play here. It seems that I've only really figure out the integer one.
Char arrays still cause the runtime to crash and float32 arrays do not give the correct results.

Let me try figuring out if there is a mismatch between the reduction on host and reduction on device.

I've noticed that accuracy is grossly sensitive to the minibatch_size which should not be happening.

This particular bug really takes the hit. It is quite hard to believe that I could screw up a single kernel this badly, but there it is.

Well, let me do the comparison and I will call it a day.

I've really made progress at a snail's pace. I can' believe it is already Sunday.