I do not feel like it anymore. Let me call it a day here.
I solved one bug, but I still have absolutely not idea what the hell is going on with that kernel.

Tomorrow I am going to insert print statements into it in order to get to the bottom of things.

As if that int thing was no confusing enough, right now I have absolutely no idea what the reduction is doing.

I rewrote it so the reduction is done on host and there accuracy goes to full 32 in 3 iterations.

Here I am getting nothing.

I need to comb it over.

This is such bullshit.

Really, there are times like this in programming. Effort will see me through.