I give. I am trying to deal with the damn memory corruption errors on the Cuda side and now string printing is fucking with me.
I should just print those turds as byte arrays on the host side and be done with it.

At any rate, in my fervor to get it over quickly, I've completely broken everything and and going to have to dedicate a few hours to clearning the mess up.

I never really had a good solution to this even last time.

...Ugggghhhh...let me get breakfast.