"1:35pm. I have no idea. I have absolutely no idea. I am even starting to suspect NVCC of doing something at this point.
I am going to have to figure out how to use cuda-gdb to get to the bottom of this.

...What if I turned off the output?
inl train_images=CudaRandom .Uniform {elem_type=float32; dim=input_size,hidden_size}
inl train_labels=CudaRandom .Uniform {elem_type=float32; dim=input_size,hidden_size}

1:40pm. First let me get rid of Mnist.
in: 0.275944, in': 0.073618
inner i = 8
in: 0.514611, in': 0.054275
inner i = 9
in: 0.500211, in': 0.488151

Unhandled Exception: ManagedCuda.CudaException: ErrorIllegalAddress: While executing a kernel, the device encountered a load or store instruction on an invalid memory address.

Yeah, I get the same error even with completely random inputs. Let me try setting the size of the inner dimension to 16.
outer i = 0
inner i = 0
in: 0.555747, in': 0.426199
inner i = 1
in: 0.463632, in': 0.560911
inner i = 2
in: 0.142320, in': 0.953480
inner i = 3
in: 0.314483, in': 0.988772
inner i = 4
in: 0.210641, in': 0.377260
inner i = 5
in: 0.357580, in': 0.170190
inner i = 6
in: 0.251366, in': 0.145745
inner i = 7
in: 0.325590, in': 0.949560
inner i = 8
in: 0.185727, in': 0.260078
inner i = 9
in: 0.126232, in': 0.640928
inner i = 10
in: 0.617376, in': 0.628088
inner i = 11
in: 0.591901, in': 0.764199
inner i = 0
in: 0.000000, in': 0.000000
inner i = 0
in: 0.000000, in': 0.000000
inner i = 0
in: 0.000000, in': 0.000000
inner i = 0

Unhandled Exception: ManagedCuda.CudaException: ErrorIllegalAddress: While executing a kernel, the device encountered a load or store instruction on an invalid memory address.

Ok, there is something strange going on here.

Let me stop it from doing any IO so I can see what the iteration is doing.

1:50pm. What if I exchange my allocator with dynamic memory allocation?

But forget that. How about I make the arduous trek to setting up the Cuda debugger?"

Yeah, I cannot solve this with my usual toolkit.

This error could be anything. I'll try setting up the debugger somehow.