I made the bug report and am thinking of what to do later. This sure put a hamper on my mood, but with the way I am compiling Cuda loops, I sort of suspected I would run into this kind of thing eventually.
Before I completely determine that NVCC is at fault though, I am going to test the standard reduce. After that, I am going to add a while loop to the language and use that instead of a tail recursive loop.

Well, let me get on with it.

For now focus on the first one.