I can confirm that that it fails even with a single tail recursive functional loop when tuples enter the picture.
I am going to try adding a while construct to the language in order to try and get around this.

I am really 90% sure now that the way NVCC is assigning to local tuples is somehow broken, but I am not sure whether the error manifests only in functional code or not, so I'll make it easier for the Cuda compiler.