Ah, fuck. I cannot put statements into the conditional in C.
F# could do that, but I do not need while in F#.

AHhhhh...I know, I know...I'll put the conditional into a join point!

This is something that I did in the very early days of Spiral for if statements. Of course, now I know the CPS trick for return insertion, but I can reuse it here. It will make for some very non-idiomatic code, but it will work.