Ok, I had enough for today. Tomorrow I will get rid of all the testing scafolding and convert the functional to imperative loops inside the kernels.
This won't be hard. I just need to clean up a bit. I really wrecked everything in the process of trying to get to the bottom of this.

But really, although it is annoying that I ran into this kind of bug with NVCC, it is hardly surprising. I am probably the first guy in the world who tried compiling kernels in this manner on the GPU.

I got some valuable debugging experience out of it, so all is well that ends well. The knowledge that I am the first one travelling this path is a valuable piece of information.

Well, since it is getting late, I am calling it a day here. There is really no getting through memory corruption bugs quickly. I should be happy it only took away two days from me.