I got up late today, but let me finally start. First off here are some optimizations.
I got rid of the check for literal in while simply because the syntax of such constructs is so fucked up in C. But C having poor syntax is hardly a surprise.

At any rate, this while loop is just a stopgap measure.

Right now the thing I have to do is to go through the kernels top to bottom and clean them up because they are litered with print statements and such. Also I need to uncomment them and get rid of the useless accuracy functions I made for debugging.

Also since I got rid of laziness from map_redo, I need to change the examples to match.