"4pm. Finally comes the learning8. Let's see if converges.
4:05pm.
-----
Batch done.
Average of batch costs is 1.67311933636665.
The accuracy of the batch is 3/32.
-----
-----
Batch done.
Average of batch costs is 0.917693890631199.
The accuracy of the batch is 4/32.
-----
-----
Batch done.
Average of batch costs is 0.708882190287113.
The accuracy of the batch is 4/32.
-----

Oh, that son of a bitch.

I guess it is time to insert print statements into the d1 kernel once again. If that causes the thing crash I am not sure what I will do."