Whatever the problem is, it is different from last time.
It is genuinely missing the correct results now.

I have no idea. I am going to have to look into one of the previous commits and get the version that reduces on host.

Or at least print out the matrices in order to make sure that they are being iterated correctly.