Ok, it works now. It really worked from the start, but the damn reshape bug got through my guard.
I should have tested for it before.

I'll remove the print statements and clean up tomorrow. Let me call it a day here.

Tomorrow I will without a doubt get the accuracy calculations to work and complete the first part of the library.

All's well that ends well and I am really glad that this has ended. This kind of debugging work is stressful and not my favorite part of programming. I still haven't gotten back from Nvidia regarding the bug report I sent, but they'll get to it eventually.

Today I fixed the serious bug in reshape and fixed the printing of C types for good.

This last one is especially good since now I have a system for doing the same in F# should I want to print struct union types without having to put everything in an `and` block.