Now the network won't train properly.
It did so last time. And using bool arrays still corrupts the stack,

I am not through the hell of debugging just yet.

Edit:

"11:50am. Is the learning rate set to low?
optimizer=Optimizer.sgd 0.5f32

Ah, yes, I can increase it by 50x and still not get divergence. I remember this vaguely.

I got fooled by the logistic regression and the previous fake result.
-----
Batch done.
Average of batch costs is 0.898405349286397.
The accuracy of the batch is 9115/60000.
-----

No, it is fucked up. It is not just the learning rate.

Just remember how long it took me to debug this shit in the original library? I am going to have to go through it all again."

I might have to look into how it was done in the old library as well so I have some idea of how valid training proceeds. It sort of faded from my memory at this point.

I was hoping for smooth progress, but it seems I am going to have to go through quite some pain to get to where I want.