Now it just optimizes on par with square. Piss poorly.
This stuff really brings me back to last 2015 when I first started doing neural nets. Those long tedious days of practice writing the nets by hand, and those vicious bugs that I could not unearth.

I got rid of them by moving to AD, based solutions, but here they are again, back to haunt me.

...

I am at my end.

Past this point it is just blind guessing. The situation I am in is trully the greatest humiliation as a programmer.

At this point the only thing I can do is make sure that all memory is safe.

Maybe as the last thing for the day, I will make a dynamic allocator and try replacing the one that I have now.

I do not feel like digging through the old library at the moment.

I really wanted to be done with this thing a lot more swiftly and move on to recurrent nets, but since it has come to this I'll forget about it for the next week or two.

Let me do the new allocator.