"5:30pm. Yeah, basically I do not feel like making a new allocator for no reason.
Well, it is not no reason – all that new Cuda stuff is dodgy as hell that is for sure. Lack of memory safety will make for all sorts of ridiculous errors that would never have manifested otherwise. Plus all that pointer alignment nonsense might lead to outright incorrect results.

As shitty as the previous memory allocation scheme was, I could count on the Cuda allocator to align the things properly. Not so here. Also all the matrices were in fact chunked rather than views in the previous library.

5:40pm. I have no choice – this is what I will aim for – make the damn trace in Spiral. Then the make the trace in the previous library. Then compare the two to see where they diverge.

This is annoying as it will be a bunch of effort I will have to throw away at the end, but I should just grit my teeth and get through it. I did it all the time in the past. This will hardly be the first time I am doing such a thing."

Honestly, I do not feel like it. It is hard to get motivated to do this sort of thing.

Regardless of what I want to do, I need to increase my test coverage. Whether it is making a new allocator or writing the AD operations for host, I'll end up doing it all in the end eventually.

So I just need to steel myself and go forward.

Let me make a trace in the previous library first.

I'll start a new project to bring it here.