I really meant to do more after the last commit instead of just adding the old library to the solution, but I guess I'll have to resume tomorrow. It is late now and I am tired.
Shameful. It is one thing to make mistakes, but to get so far behind enemy lines that the retreat is cut off is what an amateur would do.

I was too optimizstic with the regards to the library.

My plan is this, I will step by step bring it to feature parity with the old library, and will compare it every step of the way to make sure no more mistakes get made.

Tomorrow I will dedicate to making the testing harnesses, and after that I will move from there. I am also thinking of making the host versions of the functions for future safety, but I'll keep that aside.

I didn't think I'd ever touch the old library again, but it seems it will be necessary after all.

Well it is fine.

The old library is my greatest work in F#.

It is the one that really made me into what I am today, so I might as well say hello to an old friend before leaving again.

It won't be long until I've gone to a place where none may reach.

Shadow dwellers such as myself might be despited by society, but that will only remain in effect until I figure out how to convert my abilities into power.

One day Spiral will be synonymous with power. It won't be long now.