Let me start the day. I had a lot to think about last night.
From a goal based perspective going with the old library seems the most reasonable choice, but it is also the choice I like the least.

By my reasoning, these bugs should not even be happening. I specifically wrote Spiral in a high level manner in order to ensure correctness. If this was done on the CPU it would be working already - I am convinced of that.

So before I start there is a need to assign some blame to the NVCC.

Regarding the bool array bug, I am not entirely sure that it is padding them. Probably if I mapped a bool array it will work which will be proof that it is not padding it.

And even if it is padding it, that is no reason for it to corrupt the stack. Hence this is also another of NVCC's bugs.

I am going to look into seeing how to disable optimizations for it as this is ridiculous.

Also the way these network train is just so bizzare. It is like only a part of them is being trained. Hence, why don't I check out the way weights are initialized.

Usually I wouldn't need to suspect this. Yesterday I did suspect initialization for a bit, but let me print out the whole thing today just so I can take a look at it. It might turn out to be the case that a lot of the weights are getting inited to zero.

So first let me check out how large arrays are inited and then i will try mapping a char array. If it works as I suspect it will, then it will be time for another bug report.

Besides this I want to take another look in the allocator to make sure that pointers are definitely aligned.