Strangely enough this works now.
Let me run it through memcheck.

So it is not NVCC.

There is definitely some kind of memory corruption that I am not accounting for.

Alternative hypothesis:

If I assume that the nets are not training properly in the first place due to memory whackery, maybe using char arrays there is simply what triggers corruption directly?

I need to put those nets that finish running without explicit error through memcheck.

It might be the case that the functional loop thing is also related to this issue.

For now let me put this through memcheck.