I made a minimalist example where the error triggers. Since this is a memory corruption bug, I can't really pare it down properly. I notice that as I change the code, the corruption happens in completely random places and cuda-memcheck is useless since it does nothing if the program terminates. The exceptions and error messages seem only semi-deterministic as well.
I need so advice on how to deal with this. At this point I am just guessing.

I am suspecting NVCC of doing something strange as I ran into a bug with it earlier, but I can't really just send this huge example and ask them to figure it out.