I am done for the day. I found the error and submitted an issue on ManagedCuda. Let me put the join point printing on a separate branch and then I will reset tomorrow.
I am done for the day. Good job me, for finding that bug.

I am not sure whether this will be enough to get it to train properly - it might be very well the case that there are still bugs left, but resolving that will be left to the future me.

One by one, I will catch them all.