As far as I am concerned that bug is solved.
I know why it is happening period.

Now I can move on.

Note that this still does not tell me why the nets are currently training so poorly, nor why two years ago I needed to add that second synchronize in order to make sure that the transfers work. Since I am at it, I should investigate that.

Once I have that under wraps, I'll be able to figure out why the library is training so poorly with a peace of mind.