Let me resume. I have a good idea.
First I need to clean up the mess that I made yesterday. Then I will will test something. Is the cost changing if I update the weights with a learning_rate of zero? I think I tried this in the previous library once and it revealed some bugs.

There is no need to give into despair just yet.

I can get through this. Believe in the light of inspiration.