I removed the changes.
Now let me try something that I should have done before - I'll set the learning rate to zero and see whether the cost is budging.

There has to be some kind of a hint as to why it is optimizing so poorly somewhere.

Now that I hopefully have the memory errors out of the way, I'll be able to find it.

Let me just run this.