a = [[0.2925366, -0.718359, 0.09999694],[-0.3931978, 2.547023, -0.003436466]] b = [[0.008292325, -0.2509764],[0.1289789, 0.3728051],[1.082156, -0.6650119]] array([[ 0.01798494, -0.40772684],        [ 0.32153291,  1.05051182]])
Ok. Now let me make a 2x2 array so I can figure out how exactly do I need to transpose it so I can get it working right.

[|
    [|0.2925366; -0.718359|]
    [|0.09999694; -0.3931978|]
|]

[|
    [|2.547023; -0.003436466|]
    [|0.008292325; -0.2509764|]
|]

[|
    [|0.7447537; -1.828325|]
    [|-0.02267106; 0.09272648|]
|]

This is Spiral.

Now let me do the same in numpy.

>>> a
array([[ 0.2925366 , -0.718359  ],
       [ 0.09999694, -0.3931978 ]])
>>> b
array([[ 2.547023  , -0.00343647],
       [ 0.00829232, -0.2509764 ]])
>>> numpy.matmul(a,b)
array([[ 0.73914058,  0.17928586],
       [ 0.25143398,  0.09833973]])

…
>>> numpy.matmul(a.T,b.T)
array([[ 0.74475381, -0.02267106],
       [-1.82832568,  0.0927265 ]])

This is actually it, except all the fields are transposed.
>>> numpy.matmul(b,a)
array([[ 0.74475381, -1.82832568],
       [-0.02267106,  0.0927265 ]])

There it is. Now it matches Spiral.

I guess it is time that I begin pushing on Spiral's end.