I did it. It works. Now I have a basic testing harness going on.
Let me remove SpiralV4. I am truly proud that I managed to resolve this issue without resorting to comparing it side by side with another library.

Apart from the gemm thing, the memory bug and the NVCC bug which attacked me all at once and which aren't related to the architecture the system, the fact that it worked all the first time validates my approach. I can derive great deal of correctness by writing the code in a high level manner.

This is truly beyond great.

What I did in the last 4 days was some truly masterful programming.

Let me take a break here.

I need to think of what should come next anyway.

I think I was intending to do some performance tuning of kernels, but maybe I will just skip that part. It doesn't really matter what the dimension size minimums are. Hell, let me just set them to 32 and be done with it.

inl blockDimY = lit_min 32 (s dim_in_a)

Eh, it was just this one? Whatever. It is 32 now. This should increase its occupancy on Mnist and not make it much worse on anything else.

In my experiments in the previous library it did not really matter either way apart from the curious case of having a single warp per block which allowed me to do all the reductions using warp shuffles.

With this I am done for the day.

Everything in the library makes sense to me now and works as it should. I feel completely back in control. The last 4 days is like I was suddenly yanked into unfamiliar territory, but I did have faith in myself so it was not particularly stressful.

It is time I start making plans for the next step. Is it time for RNNs and poker? It could be.