I added the `Known Bugs` section in the documentation and put in nan's and blittable types.
Actually, it seems a lot of part 2 is actually new to me.

Change of plans – I am going to have to dedicate a day or two for the sake of going through that stuff after all.

Part 1 is stuff I already knew before, but this is very cutting edge.

Let me call it a day. For not doing anything productive all day, I sure am dead tired. It is no wonder.

Forming motivations takes a lot of energy and so does learning new things.

I won't allow myself to linger too much on this.

I'll try to get all 7 videos out of the way in the next two days.