http://course.fast.ai/lessons/lessons2.html
Here is the final plan. I will skip the first half of part two and watch the stuff on attention models and dense nets tomorrow as that might turn out to actually be useful.

The stuff on GANs is just a gimic that I do not need to burden myself with.

This is about 6h of videos to go through, but investing one day into this might prove useful down the road.

Just now I've been struck on how trivial what I really want to do is.

Yesterday it felt like there is a huge amount of work ahead of me, but now that I really look more carefully, I can see that I am pretty much done for I have pretty much trivialized neural nets with Spiral.

I'll be done in about two weeks for good and then it is time for me to just go and pick up the money. Now that will actually be a lot of work, but with my soul at ease at having ventured through the passage, I will be able to win.

I am done.

As basic as what I have now seems, I've pretty much done 80% of the difficult stuff.

It is ridiculous.

I've made a language and I feel that in terms of programming ability there probably aren't even 100 people in the world that can measure up to my current level.

If what I am saying is true, then there really should not be any obstacles as I move forward like last year.

The thing that comes to mind to me is whether it really was necessary to go through all this lengths to making a language? I mean I could have just used PyTorch or something else, couldn't I?

I can't deny that no sane person would go to these lengths just to do something as simple as I just did.

But looking at myself I can't feel any shame nor regret.

The thing might be simple and trivial, and those in the future might not understand what the big deal about this was nor about Spiral.

But I was the one who made it simple. I was the one who made it trivial.

This is my victory.